Concluído versão 1.0

+++ src/component/TabelaContato.js
@@ -33,16 +33,16 @@ export default class TabelaContatoComponent extends Component {
                             {_this.props.itens.map(function (item, index) {
                                 return (
                                     <tr key={index}>
-                                        <td>{item['.key']}</td>
+                                        <td>{item.id}</td>
                                         <td>{item.nome}</td>
-                                        <td>{item.ramal}</td>
+                                        <td>{item.ramalOuTelefone}</td>
                                         <td>{item.setor}</td>
                                         <td>
                                             <a className="button is-primary" onClick={() => {_this.onEdit(item)}}>
                                                 <span className="icon is-small">
                                                     <i className="fa fa-pencil-square-o"></i>
                                                 </span>
-                                            </a> <a onClick={() => _this.onRemove(item)} className="button is-danger">
+                                            </a> <a onClick={() => _this.onRemove(item, index)} className="button is-danger">
                                                 <span className="icon is-small">
                                                     <i className="fa fa-trash"></i>
                                                 </span>
@@ -63,15 +63,20 @@ export default class TabelaContatoComponent extends Component {
         this.props.onSearch(this.refs.filter.value.trim());
     }
 
-    onRemove = (item) => {
-        this.props.onRemove(item);
+    onRemove = (item, index) => {
+        let remove = confirm('Tem certeza que deseja remover esse contato?');
+        if (remove) {
+            this.props.itens.splice(index, 1);
+            this.props.onRemove(item);
+        }
+        
     }
 
     onEdit = (item) => {
         this.props.onEdit({
-            id: item['.key'],
+            id: item.id,
             nome: item.nome,
-            ramal: item.ramal,
+            ramalOuTelefone: item.ramalOuTelefone,
             setor: item.setor
         });
     }
